Fix testAppiumOptions tests.
